Refactor: extract inflation code to inflate module
