+++ lib/write.js
@@ -1,8 +1,8 @@
 const b4a = require('b4a')
-const c = require('compact-encoding')
 const { OP_COHORT } = require('./compression.js')
 const { encodeBlock, TYPE_COMPAT, TYPE_LATEST } = require('./encoding.js')
 const { Pointer, KeyPointer, ValuePointer, TreeNode, INSERTED, NEEDS_SPLIT } = require('./tree.js')
+const { inflateValue } = require('./inflate.js')
 
 const PREFERRED_BLOCK_SIZE = 4096
 const INLINE_VALUE_SIZE = 1024
@@ -128,7 +128,7 @@ module.exports = class WriteBatch {
         c = b4a.compare(target, m.key)
 
         if (c === 0) {
-          const existing = await snap.inflateValue(m, conf)
+          const existing = await inflateValue(m, conf)
           if (b4a.equals(existing, value)) return false
           v.setValue(this.tree.context, mid, value)
           for (let i = 0; i < stack.length; i++) stack[i].changed = true
@@ -149,7 +149,7 @@ module.exports = class WriteBatch {
     if (status >= 0) {
       // already exists, upsert if changed
       const m = v.keys.uget(status)
-      const existing = await snap.inflateValue(m, conf)
+      const existing = await inflateValue(m, conf)
       if (b4a.equals(existing, value)) return false
       v.setValue(this.tree.context, status, value)
     }
